PB/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
dj-geocoding==0.2.1
Django==4.1.3
django-cors-headers==3.13.0
django-geoposition==0.3.0
django-jsonfield==1.4.1
django-location-field==2.1.0
django-mapbox-location-field==2.0.0
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
geographiclib==1.52
geopy==2.2.0
googlemaps==4.6.0
idna==3.4
Pillow==9.3.0
pygeocodio==1.1.0
PyJWT==1.7.1
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
